requirements/production.txt
django-markdownify-filter @ https://github.com/johannesdewit/django-markdownify-filter/tarball/master
betacode==1.0
pygtrie
Django==4.*
django-fullurl
gunicorn==20.*
whitenoise==5.*
PyYAML==6.*
requests==2.*
click==8.*
tqdm==4.*
regex
Unidecode==1.*
django-extensions==3.*
fuzzysearch==0.7.3
Markdown==3.3.6
pyCTS==0.2.2
beautifulsoup4==4.11.1
aenum==3.*
django-compressor==4.*
django-libsass==0.*
python-decouple==3.*